Lumina/requirements.txt
# Lumina - AI Creative Production
# Core dependencies
langgraph>=0.4.0
langchain-core>=0.3.0
langchain-openai>=0.3.0
langchain-anthropic>=0.3.0

# API
fastapi>=0.115.0
uvicorn>=0.34.0
httpx>=0.28.0

# Database
supabase>=2.0.0
asyncpg>=0.30.0
pgvector>=0.3.0

# Image/Video processing
Pillow>=11.0.0

# Data
pydantic>=2.10.0
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0

# Type checking
mypy>=1.14.0
